# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealthDesk Plus API",
        "description": "Comprehensive healthcare management system for doctors, patients, appointments, prescriptions, and medical records.",
        "contact": {
            "name": "HealthDesk Plus Support",
            "email": "support@healthdeskplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.healthdeskplus.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/ai/process-document": {
            "post": {
                "tags": [
                    "Document Processing"
                ],
                "summary": "Process uploaded document with OCR and AI extraction",
                "description": "Upload a document and extract structured data using OCR and AI",
                "operationId": "16fa71686a8e53c4766c7a85690cc391",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "document": {
                                        "description": "Document file (image or PDF)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "document_type": {
                                        "description": "Type of document",
                                        "type": "string",
                                        "enum": [
                                            "certificate",
                                            "license",
                                            "id_card",
                                            "degree"
                                        ],
                                        "example": "certificate"
                                    },
                                    "auto_fill_form": {
                                        "description": "Whether to auto-fill form fields",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "document_type": {
                                                    "type": "string",
                                                    "example": "certificate"
                                                },
                                                "raw_text": {
                                                    "type": "string",
                                                    "example": "Dr. John Doe MBBS..."
                                                },
                                                "structured_data": {
                                                    "properties": {
                                                        "doctor_name": {
                                                            "type": "string",
                                                            "example": "Dr. John Doe"
                                                        },
                                                        "registration_number": {
                                                            "type": "string",
                                                            "example": "ABC123456"
                                                        },
                                                        "qualification": {
                                                            "type": "string",
                                                            "example": "MBBS, MD"
                                                        },
                                                        "issuing_authority": {
                                                            "type": "string",
                                                            "example": "Medical Council of India"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "confidence": {
                                                    "type": "number",
                                                    "example": 0.85
                                                },
                                                "validation": {
                                                    "properties": {
                                                        "is_valid": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "confidence_score": {
                                                            "type": "number",
                                                            "example": 0.85
                                                        },
                                                        "missing_fields": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "warnings": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "suggestions": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "form_data": {
                                                    "description": "Auto-filled form data",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ai/batch-process-documents": {
            "post": {
                "tags": [
                    "Document Processing"
                ],
                "summary": "Process multiple documents in batch",
                "description": "Upload multiple documents and extract structured data from all",
                "operationId": "2a6e2af2634e39e02c5bbc2b8490dfdd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "documents": {
                                        "description": "Document files",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "document_types": {
                                        "description": "Types of documents (certificate, license, etc.)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documents processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "document_type": {
                                                        "type": "string"
                                                    },
                                                    "structured_data": {
                                                        "type": "object"
                                                    },
                                                    "confidence": {
                                                        "type": "number"
                                                    },
                                                    "validation": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ai/extract-text": {
            "post": {
                "tags": [
                    "Document Processing"
                ],
                "summary": "Extract text from document using OCR",
                "description": "Extract raw text from uploaded document using OCR",
                "operationId": "644f1ecbebaeab1df44900085be3584f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "document": {
                                        "description": "Document file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Text extracted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "text": {
                                                    "type": "string",
                                                    "example": "Dr. John Doe MBBS..."
                                                },
                                                "confidence": {
                                                    "type": "number",
                                                    "example": 0.85
                                                },
                                                "word_count": {
                                                    "type": "integer",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}